feat(storage): add hability to specify a Buffer to Writer
In order to reduce the number of allocations when uploading files the
the user can supply it's own buffer to be used instead of allocating
a new one each time. The user needs to keep track of this buffer
(e.g with sync.Pool) and ensure the buffer is not used in multiple threads
otherwise a race condition may occur.